Move ThreadContext out of io-context.h
Just breaking things up a bit to make io-context.t smaller.
