+++ src/workerd/io/io-context.c++
@@ -125,29 +125,6 @@ public:
   kj::Maybe<kj::Promise<void>> maybePromise;
 };
 
-ThreadContext::HeaderIdBundle::HeaderIdBundle(kj::HttpHeaderTable::Builder& builder)
-    : table(builder.getFutureTable()),
-      contentEncoding(builder.add("Content-Encoding")),
-      cfCacheStatus(builder.add("CF-Cache-Status")),
-      cacheControl(builder.add("Cache-Control")),
-      cfCacheNamespace(builder.add("CF-Cache-Namespace")),
-      cfKvMetadata(builder.add("CF-KV-Metadata")),
-      cfR2ErrorHeader(builder.add("CF-R2-Error")),
-      cfBlobMetadataSize(builder.add("CF-R2-Metadata-Size")),
-      cfBlobRequest(builder.add("CF-R2-Request")),
-      authorization(builder.add("Authorization")),
-      secWebSocketProtocol(builder.add("Sec-WebSocket-Protocol")) {}
-
-ThreadContext::ThreadContext(
-    kj::Timer& timer, kj::EntropySource& entropySource,
-    HeaderIdBundle headerIds, capnp::HttpOverCapnpFactory& httpOverCapnpFactory, capnp::ByteStreamFactory& byteStreamFactory, bool fiddle)
-    : timer(timer),
-      entropySource(entropySource),
-      headerIds(headerIds),
-      httpOverCapnpFactory(httpOverCapnpFactory),
-      byteStreamFactory(byteStreamFactory),
-      fiddle(fiddle) {}
-
 IoContext::IoContext(ThreadContext& thread,
                                kj::Own<const Worker> workerParam,
                                kj::Maybe<Worker::Actor&> actorParam,
